Updated s3 benchmark

+++ src/s3.cpp
@@ -125,16 +125,18 @@ void s3::empty_bucket(const Aws::String &bucket_name) {
   ListObjectsRequest list_request;
   list_request.SetBucket(bucket_name);
 
-  ListObjectsOutcome list_outcome = m_client->ListObjects(list_request);
+  while (true) {
+    ListObjectsOutcome list_outcome = m_client->ListObjects(list_request);
 
-  if (!list_outcome.IsSuccess())
-    return;
+    if (!list_outcome.IsSuccess() || list_outcome.GetResult().GetContents().empty())
+      return;
 
-  for (const auto &object : list_outcome.GetResult().GetContents()) {
-    DeleteObjectRequest delete_request;
-    delete_request.SetBucket(bucket_name);
-    delete_request.SetKey(object.GetKey());
-    m_client->DeleteObject(delete_request);
+    for (const auto &object : list_outcome.GetResult().GetContents()) {
+      DeleteObjectRequest delete_request;
+      delete_request.SetBucket(bucket_name);
+      delete_request.SetKey(object.GetKey());
+      m_client->DeleteObject(delete_request);
+    }
   }
 }
 void s3::wait_for_bucket_to_empty(const Aws::String &bucket_name) {
